Keep runtime capability modularization scoped to the current loong surface
The original PR branch was carrying unrelated daemon churn from a stale base and
reintroduced loongclaw naming that no longer matches the repository. This rewrite
rebuilds the runtime capability CLI split directly on top of current dev, keeps
the existing command/render contract, and moves the internal helpers into focused
submodules without dragging in the unrelated refactors.

A small session wait test tolerance adjustment is included because the full
workspace gate exposed a timing-sensitive assertion under suite load after the
refactor verification pass.

Constraint: PR 1314 must stay mergeable on the current loong-branded dev branch
Rejected: Merge the original PR as-is | unrelated stale-base refactors and branding regressions made the diff unsafe
Rejected: Port the broader daemon reshuffle too | issue #1309 only calls for runtime capability CLI cleanup
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep runtime capability entrypoint behavior stable; future cleanup should stay inside the runtime capability module unless a separate issue expands scope
Tested: ./scripts/cargo-local-toolchain.sh fmt --all -- --check
Tested: ./scripts/cargo-local-toolchain.sh clippy --workspace --all-targets --all-features -- -D warnings
Tested: ./scripts/cargo-local-toolchain.sh test -p loong --test integration runtime_capability_ -- --nocapture
Tested: ./scripts/cargo-local-toolchain.sh test -p loong-app --lib session_wait_wakes_when_parent_mailbox_receives_delegate_result -- --nocapture
Tested: ./scripts/cargo-local-toolchain.sh test --workspace
Tested: ./scripts/cargo-local-toolchain.sh test --workspace --all-features
Not-tested: GitHub Actions rerun on the updated PR branch
