+++ crates/app/src/tools/session.rs
@@ -7583,6 +7583,10 @@ mod tests {
             assert!(send_result.is_ok());
         });
 
+        let wait_timeout_ms = 1_000_u64;
+        let poll_interval_ms = 10_usize;
+        let early_wake_budget_ms = 100_u64;
+        let expected_max_wait = Duration::from_millis(wait_timeout_ms - early_wake_budget_ms);
         let started_at = Instant::now();
         let outcome = wait_for_single_session_with_policies(
             "child-session",
@@ -7590,8 +7594,8 @@ mod tests {
             &config,
             &ToolConfig::default(),
             None,
-            1_000,
-            10,
+            wait_timeout_ms,
+            poll_interval_ms,
         )
         .await
         .expect("session_wait outcome");
@@ -7600,7 +7604,7 @@ mod tests {
         assert_eq!(outcome.status, "ok");
         assert_eq!(outcome.payload["wait_status"], "completed");
         assert_eq!(outcome.payload["session"]["state"], "completed");
-        assert!(started_at.elapsed() < Duration::from_millis(500));
+        assert!(started_at.elapsed() < expected_max_wait);
     }
 
     #[test]
